Revert "use int64 as vout"
This reverts commit c05e200e6bd04702e353ca54abc478b1433fd6ca.

+++ api/types.go
@@ -185,7 +185,7 @@ const (
 
 // AddressFilter is used to filter data returned from GetAddress api method
 type AddressFilter struct {
-	Vout           int64
+	Vout           int
 	Contract       string
 	FromHeight     uint32
 	ToHeight       uint32
